Implement q_ascend and q_descend function
Add the q_ascend function that can
remove every node which has a node with
a strictly less value anywhere to the right side of it.

Add the q_descend function that can
remove every node which has a node with
a strictly greater value anywhere to the right side of it.

Change-Id: I874bbd197dacb26b21043272c8c4527ee011bef1